# Move flattening/unflattening logic of JIT to C