@apaszke 2017-11-04
Yes, in `benchmark/scripts/lstm.py` with flag `--jit` the time drops from around 204us to 164us. Time spent on flattening/building up in_key drops from >40us to 6us. Unflattening is slightly faster, but it didn't seem to be such a significant bottleneck before.

I'll add the comments later