@zdevito 2017-11-07
Looks generally good, with a great improvement to building in_key. I couldn't find where the tensor sizes are encoded in the key in the new version. Are we no longer using that info or am I just missing the line that does it?