@apaszke 2017-11-07
Yeah I just realized that we no longer check the sizes and strides. I'll fix that in a PR that moves more of tracing logic to C++