@dzhulgakov 2017-11-14
Hey folks, just breaking batch_norm export (and disabling tests) is not cool as it completely breaks ONNX export for anything having batchnorm!

In fact, since we're doing match by the number of inputs when we're exporting the initializers in the graph (https://github.com/pytorch/pytorch/blob/master/torch/csrc/jit/export.cpp#L221) I suspect it also breaks any export if some of the modules has any buffers (since we compare the order of state_dict and parameters())

Is there a more plausible fix for buffers handling? One option might be to just override the symbolic export for the entire module (batchnorm) that has buffers but that would require hack similar to https://github.com/pytorch/pytorch/pull/3044/

@apaszke @ezyang @houseroad - any suggestions?