@apaszke 2017-11-14
Oops, my bad. I didn't break the onnx pass on purpose, the test I disabled was meant to be a JIT check, and we don't intend to be able to re-execute batch norm ops yet. I'll add back support for tensors today in the evening.