@apaszke 2017-11-14
(This will only be a hotfix. We should figure out a more robust solution.)