@apaszke 2017-11-14
Actually, once Variables and Tensors will be fused, we won't have a clear distinction between parameters and buffers and this will simplify the code a lot (e.g. BatchNorm will be able to take buffers as inputs)