@apaszke 2017-11-14
Fixed in #3691